+++ CHANGELOG.md
@@ -6,7 +6,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/)
 
 
-## [In Development] - Unreleased
+## [1.11.1] - 2022-02-25
+
+### Added
+
+- More logging to the Discord message logic
+
+### Fixed
+
+- Enable editable SRP amount field only for pending and rejected requests (#51)
+- CSS classes in templates
+- Only count approved SRP requests towards the total SRP costs
+
+### Changed
+
+- JavaScript modernised
 
 
 ## [1.11.0] - 2022-02-02